db: user & session
Adds Prisma as a dependency and sets up user & session tracking in the database. Uses Postgres, a dev database is included in docker-compose.yml.

Users will be authenticated through Plex's OAuth API. Sessions will be tracked with a JWT on the client side which maps to a unique Plex auth token per session.