feat: auth helpers
Implements helpers for getting the server secret from the environment, and checking if the user is authenticated (in a lightweight fashion, it doesn't check the session against the database as I intend to use the helper in middleware).

Session verification will be left to controller logic that actually changes database state.